# convert to using generics