--- a/cache.go
+++ b/cache.go
@@ -61,7 +61,7 @@ const (
 )
 
 // Config configures the cache.
-type Config struct {
+type Config[K comparable, V any] struct {
 	// Maximum number of items in the cache
 	Capacity int
 	// Optional max duration before an item expires. Must be greater than or
@@ -77,28 +77,28 @@ type Config struct {
 	// to the MaxAge
 	ExpirationInterval time.Duration
 	// Optional callback invoked when an item is evicted due to the LRU policy
-	OnEviction func(key, value interface{})
+	OnEviction func(key K, value V)
 	// Optional callback invoked when an item expired
-	OnExpiration func(key, value interface{})
+	OnExpiration func(key K, value V)
 }
 
 // Entry pointed to by each list.Element
-type cacheEntry struct {
-	key       interface{}
-	value     interface{}
+type cacheEntry[K comparable, V any] struct {
+	key       K
+	value     V
 	timestamp time.Time
 }
 
 // Cache implements a thread-safe fixed-capacity LRU cache.
-type Cache struct {
+type Cache[K comparable, V any] struct {
 	// Fields defined by configuration
 	capacity           int
 	minAge             time.Duration
 	maxAge             time.Duration
 	expirationType     ExpirationType
 	expirationInterval time.Duration
-	onEviction         func(key, value interface{})
-	onExpiration       func(key, value interface{})
+	onEviction         func(key K, value V)
+	onExpiration       func(key K, value V)
 
 	// Cache statistics
 	sets      int64
@@ -107,7 +107,7 @@ type Cache struct {
 	misses    int64
 	evictions int64
 
-	items        map[interface{}]*list.Element
+	items        map[K]*list.Element
 	evictionList *list.List
 	mutex        sync.RWMutex
 	rand         RandGenerator
@@ -117,7 +117,7 @@ type Cache struct {
 // must be a positive int, and config.MaxAge a zero or positive duration. A
 // duration of zero disables item expiration. Panics given an invalid
 // config.Capacity or config.MaxAge.
-func New(config Config) *Cache {
+func New[K comparable, V any](config Config[K, V]) *Cache[K, V] {
 	if config.Capacity <= 0 {
 		panic("Must supply a positive config.Capacity")
 	}
@@ -146,7 +146,7 @@ func New(config Config) *Cache {
 
 	seed := rand.NewSource(time.Now().UnixNano())
 
-	cache := &Cache{
+	cache := &Cache[K, V]{
 		capacity:           config.Capacity,
 		maxAge:             config.MaxAge,
 		minAge:             minAge,
@@ -154,7 +154,7 @@ func New(config Config) *Cache {
 		expirationInterval: interval,
 		onEviction:         config.OnEviction,
 		onExpiration:       config.OnExpiration,
-		items:              make(map[interface{}]*list.Element),
+		items:              make(map[K]*list.Element),
 		evictionList:       list.New(),
 		rand:               rand.New(seed),
 	}
@@ -171,8 +171,8 @@ func New(config Config) *Cache {
 }
 
 // Set updates a key:value pair in the cache. Returns true if an eviction
-// occurrred, and subsequently invokes the OnEviction callback.
-func (cache *Cache) Set(key, value interface{}) bool {
+// occurred, and subsequently invokes the OnEviction callback.
+func (cache *Cache[K, V]) Set(key K, value V) bool {
 	cache.mutex.Lock()
 	defer cache.mutex.Unlock()
 
@@ -181,13 +181,13 @@ func (cache *Cache) Set(key, value interface{}) bool {
 
 	if element, ok := cache.items[key]; ok {
 		cache.evictionList.MoveToFront(element)
-		entry := element.Value.(*cacheEntry)
+		entry := element.Value.(*cacheEntry[K, V])
 		entry.value = value
 		entry.timestamp = timestamp
 		return false
 	}
 
-	entry := &cacheEntry{key, value, timestamp}
+	entry := &cacheEntry[K, V]{key, value, timestamp}
 	element := cache.evictionList.PushFront(entry)
 	cache.items[key] = element
 
@@ -198,17 +198,17 @@ func (cache *Cache) Set(key, value interface{}) bool {
 	return evict
 }
 
-// Get returns the value stored at `key`. The boolean value reports whether or
-// not the value was found. The OnExpiration callback is invoked if the value
+// Get returns the value stored at `key`. The boolean value reports whether
+//  the value was found. The OnExpiration callback is invoked if the value
 // had expired on access
-func (cache *Cache) Get(key interface{}) (interface{}, bool) {
+func (cache *Cache[K, V]) Get(key K) (value V, found bool) {
 	cache.mutex.Lock()
 	defer cache.mutex.Unlock()
 
 	cache.gets++
 
 	if element, ok := cache.items[key]; ok {
-		entry := element.Value.(*cacheEntry)
+		entry := element.Value.(*cacheEntry[K, V])
 		if cache.maxAge == 0 || time.Since(entry.timestamp) <= cache.maxAge {
 			cache.evictionList.MoveToFront(element)
 			cache.hits++
@@ -221,16 +221,16 @@ func (cache *Cache) Get(key interface{}) (interface{}, bool) {
 		if cache.onExpiration != nil {
 			cache.onExpiration(entry.key, entry.value)
 		}
-		return nil, false
+		return value, false
 	}
 
 	cache.misses++
-	return nil, false
+	return value, false
 }
 
-// Has returns whether or not the `key` is in the cache without updating
+// Has returns whether the `key` is in the cache without updating
 // how recently it was accessed or deleting it for having expired.
-func (cache *Cache) Has(key interface{}) bool {
+func (cache *Cache[K, V]) Has(key K) bool {
 	cache.mutex.RLock()
 	defer cache.mutex.RUnlock()
 
@@ -239,22 +239,22 @@ func (cache *Cache) Has(key interface{}) bool {
 }
 
 // Peek returns the value at the specified key and a boolean specifying whether
-// or not it was found, without updating how recently it was accessed or
+//  it was found, without updating how recently it was accessed or
 // deleting it for having expired.
-func (cache *Cache) Peek(key interface{}) (interface{}, bool) {
+func (cache *Cache[K, V]) Peek(key K) (value V, found bool) {
 	cache.mutex.RLock()
 	defer cache.mutex.RUnlock()
 
 	if element, ok := cache.items[key]; ok {
-		return element.Value.(*cacheEntry).value, true
+		return element.Value.(*cacheEntry[K, V]).value, true
 	}
 
-	return nil, false
+	return value, false
 }
 
 // Remove removes the provided key from the cache, returning a bool indicating
-// whether or not it existed.
-func (cache *Cache) Remove(key interface{}) bool {
+// whether it existed.
+func (cache *Cache[K, V]) Remove(key K) bool {
 	cache.mutex.Lock()
 	defer cache.mutex.Unlock()
 
@@ -269,7 +269,7 @@ func (cache *Cache) Remove(key interface{}) bool {
 // EvictOldest removes the oldest item from the cache, while also invoking any
 // eviction callback. A bool is returned indicating whether or not an item was
 // removed
-func (cache *Cache) EvictOldest() bool {
+func (cache *Cache[K, V]) EvictOldest() bool {
 	cache.mutex.Lock()
 	defer cache.mutex.Unlock()
 
@@ -277,7 +277,7 @@ func (cache *Cache) EvictOldest() bool {
 }
 
 // Len returns the number of items in the cache.
-func (cache *Cache) Len() int {
+func (cache *Cache[K, V]) Len() int {
 	cache.mutex.RLock()
 	defer cache.mutex.RUnlock()
 
@@ -285,7 +285,7 @@ func (cache *Cache) Len() int {
 }
 
 // Clear empties the cache.
-func (cache *Cache) Clear() {
+func (cache *Cache[K, V]) Clear() {
 	cache.mutex.Lock()
 	defer cache.mutex.Unlock()
 
@@ -296,11 +296,11 @@ func (cache *Cache) Clear() {
 }
 
 // Keys returns all keys in the cache.
-func (cache *Cache) Keys() []interface{} {
+func (cache *Cache[K, V]) Keys() []K {
 	cache.mutex.RLock()
 	defer cache.mutex.RUnlock()
 
-	keys := make([]interface{}, len(cache.items))
+	keys := make([]K, len(cache.items))
 	i := 0
 
 	for key := range cache.items {
@@ -312,15 +312,15 @@ func (cache *Cache) Keys() []interface{} {
 }
 
 // OrderedKeys returns all keys in the cache, ordered from oldest to newest.
-func (cache *Cache) OrderedKeys() []interface{} {
+func (cache *Cache[K, V]) OrderedKeys() []K {
 	cache.mutex.RLock()
 	defer cache.mutex.RUnlock()
 
-	keys := make([]interface{}, len(cache.items))
+	keys := make([]K, len(cache.items))
 	i := 0
 
 	for element := cache.evictionList.Back(); element != nil; element = element.Prev() {
-		keys[i] = element.Value.(*cacheEntry).key
+		keys[i] = element.Value.(*cacheEntry[K, V]).key
 		i++
 	}
 
@@ -330,7 +330,7 @@ func (cache *Cache) OrderedKeys() []interface{} {
 // SetMaxAge updates the max age for items in the cache. A duration of zero
 // disables expiration. A negative duration, or one that is less than minAge,
 // results in an error.
-func (cache *Cache) SetMaxAge(maxAge time.Duration) error {
+func (cache *Cache[K, V]) SetMaxAge(maxAge time.Duration) error {
 	if maxAge < 0 {
 		return errors.New("Must supply a zero or positive maxAge")
 	} else if maxAge < cache.minAge {
@@ -348,7 +348,7 @@ func (cache *Cache) SetMaxAge(maxAge time.Duration) error {
 // SetMinAge updates the min age for items in the cache. A duration of zero
 // or equal to maxAge disables jitter. A negative duration, or one that is
 // greater than maxAge, results in an error.
-func (cache *Cache) SetMinAge(minAge time.Duration) error {
+func (cache *Cache[K, V]) SetMinAge(minAge time.Duration) error {
 	if minAge < 0 {
 		return errors.New("Must supply a zero or positive minAge")
 	} else if minAge > cache.maxAge {
@@ -368,7 +368,7 @@ func (cache *Cache) SetMinAge(minAge time.Duration) error {
 }
 
 // OnEviction sets the eviction callback.
-func (cache *Cache) OnEviction(callback func(key, value interface{})) {
+func (cache *Cache[K, V]) OnEviction(callback func(key K, value V)) {
 	cache.mutex.Lock()
 	defer cache.mutex.Unlock()
 
@@ -376,7 +376,7 @@ func (cache *Cache) OnEviction(callback func(key, value interface{})) {
 }
 
 // OnExpiration sets the expiration callback.
-func (cache *Cache) OnExpiration(callback func(key, value interface{})) {
+func (cache *Cache[K, V]) OnExpiration(callback func(key K, value V)) {
 	cache.mutex.Lock()
 	defer cache.mutex.Unlock()
 
@@ -384,7 +384,7 @@ func (cache *Cache) OnExpiration(callback func(key, value interface{})) {
 }
 
 // Stats returns cache stats.
-func (cache *Cache) Stats() Stats {
+func (cache *Cache[K, V]) Stats() Stats {
 	cache.mutex.RLock()
 	defer cache.mutex.RUnlock()
 
@@ -401,7 +401,7 @@ func (cache *Cache) Stats() Stats {
 
 // Resize the cache to hold at most n entries. If n is smaller than the current
 // size, entries are evicted to fit the new size. It errors if n <= 0.
-func (cache *Cache) Resize(n int) error {
+func (cache *Cache[K, V]) Resize(n int) error {
 	if n <= 0 {
 		return errors.New("must supply a positive capacity to Resize")
 	}
@@ -421,14 +421,14 @@ func (cache *Cache) Resize(n int) error {
 	return nil
 }
 
-func (cache *Cache) deleteExpired() {
+func (cache *Cache[K, V]) deleteExpired() {
 	keys := cache.Keys()
 
 	for i := range keys {
 		cache.mutex.Lock()
 
 		if element, ok := cache.items[keys[i]]; ok {
-			entry := element.Value.(*cacheEntry)
+			entry := element.Value.(*cacheEntry[K, V])
 			if cache.maxAge > 0 && time.Since(entry.timestamp) > cache.maxAge {
 				cache.deleteElement(element)
 				if cache.onExpiration != nil {
@@ -441,7 +441,7 @@ func (cache *Cache) deleteExpired() {
 	}
 }
 
-func (cache *Cache) evictOldest() bool {
+func (cache *Cache[K, V]) evictOldest() bool {
 	element := cache.evictionList.Back()
 	if element == nil {
 		return false
@@ -455,14 +455,14 @@ func (cache *Cache) evictOldest() bool {
 	return true
 }
 
-func (cache *Cache) deleteElement(element *list.Element) *cacheEntry {
+func (cache *Cache[K, V]) deleteElement(element *list.Element) *cacheEntry[K, V] {
 	cache.evictionList.Remove(element)
-	entry := element.Value.(*cacheEntry)
+	entry := element.Value.(*cacheEntry[K, V])
 	delete(cache.items, entry.key)
 	return entry
 }
 
-func (cache *Cache) getTimestamp() time.Time {
+func (cache *Cache[K, V]) getTimestamp() time.Time {
 	timestamp := time.Now()
 	if cache.minAge == cache.maxAge {
 		return timestamp