--- a/cache_test.go
+++ b/cache_test.go
@@ -10,23 +10,23 @@ import (
 
 func TestInvalidCapacity(t *testing.T) {
 	assert.Panics(t, func() {
-		New(Config{Capacity: 0})
+		New(Config[string, int]{Capacity: 0})
 	})
 }
 
 func TestInvalidMaxAge(t *testing.T) {
 	assert.Panics(t, func() {
-		New(Config{Capacity: 1, MaxAge: -1 * time.Hour})
+		New(Config[string, int]{Capacity: 1, MaxAge: -1 * time.Hour})
 	})
 }
 
 func TestInvalidMinAge(t *testing.T) {
 	assert.Panics(t, func() {
-		New(Config{Capacity: 1, MinAge: -1 * time.Hour})
+		New(Config[string, int]{Capacity: 1, MinAge: -1 * time.Hour})
 	})
 
 	assert.Panics(t, func() {
-		New(Config{
+		New(Config[string, int]{
 			Capacity: 1,
 			MaxAge:   time.Hour,
 			MinAge:   2 * time.Hour,
@@ -35,7 +35,7 @@ func TestInvalidMinAge(t *testing.T) {
 }
 
 func TestBasicSetGet(t *testing.T) {
-	cache := New(Config{Capacity: 2})
+	cache := New(Config[string, int]{Capacity: 2})
 	cache.Set("foo", 1)
 	cache.Set("bar", 2)
 
@@ -49,7 +49,7 @@ func TestBasicSetGet(t *testing.T) {
 }
 
 func TestBasicSetOverwrite(t *testing.T) {
-	cache := New(Config{Capacity: 2})
+	cache := New(Config[string, int]{Capacity: 2})
 	cache.Set("foo", 1)
 	evict := cache.Set("foo", 2)
 	val, ok := cache.Get("foo")
@@ -60,11 +60,12 @@ func TestBasicSetOverwrite(t *testing.T) {
 }
 
 func TestEviction(t *testing.T) {
-	var k, v interface{}
+	var k string
+	var v int
 
-	cache := New(Config{
+	cache := New(Config[string, int]{
 		Capacity: 2,
-		OnEviction: func(key, value interface{}) {
+		OnEviction: func(key string, value int) {
 			k = key
 			v = value
 		},
@@ -77,23 +78,24 @@ func TestEviction(t *testing.T) {
 
 	assert.True(t, evict)
 	assert.False(t, ok)
-	assert.Nil(t, val)
+	assert.Zero(t, val)
 	assert.Equal(t, "foo", k)
 	assert.Equal(t, 1, v)
 }
 
 func TestExpiration(t *testing.T) {
-	var k, v interface{}
+	var k string
+	var v int
 	var eviction bool
 
-	cache := New(Config{
+	cache := New(Config[string, int]{
 		Capacity: 1,
 		MaxAge:   time.Millisecond,
-		OnExpiration: func(key, value interface{}) {
+		OnExpiration: func(key string, value int) {
 			k = key
 			v = value
 		},
-		OnEviction: func(key, value interface{}) {
+		OnEviction: func(key string, value int) {
 			eviction = true
 		},
 	})
@@ -103,7 +105,7 @@ func TestExpiration(t *testing.T) {
 
 	val, ok := cache.Get("foo")
 	assert.False(t, ok)
-	assert.Nil(t, val)
+	assert.Zero(t, val)
 	assert.Equal(t, "foo", k)
 	assert.Equal(t, 1, v)
 	assert.False(t, eviction)
@@ -125,7 +127,7 @@ func (mock *MockRandGenerator) Int63n(n int64) int64 {
 }
 
 func TestJitter(t *testing.T) {
-	cache := New(Config{
+	cache := New(Config[string, string]{
 		Capacity: 1,
 		MaxAge:   350 * time.Millisecond,
 		MinAge:   time.Millisecond,
@@ -165,7 +167,7 @@ func TestJitter(t *testing.T) {
 }
 
 func TestHas(t *testing.T) {
-	cache := New(Config{Capacity: 1, MaxAge: time.Millisecond})
+	cache := New(Config[string, string]{Capacity: 1, MaxAge: time.Millisecond})
 	cache.Set("foo", "bar")
 	<-time.After(time.Millisecond * 2)
 
@@ -174,7 +176,7 @@ func TestHas(t *testing.T) {
 }
 
 func TestPeek(t *testing.T) {
-	cache := New(Config{Capacity: 1, MaxAge: time.Millisecond})
+	cache := New(Config[string, string]{Capacity: 1, MaxAge: time.Millisecond})
 	cache.Set("foo", "bar")
 	<-time.After(time.Millisecond * 2)
 
@@ -186,9 +188,9 @@ func TestPeek(t *testing.T) {
 func TestRemove(t *testing.T) {
 	var eviction bool
 
-	cache := New(Config{
+	cache := New(Config[string, string]{
 		Capacity: 1,
-		OnEviction: func(key, value interface{}) {
+		OnEviction: func(key, value string) {
 			eviction = true
 		},
 	})
@@ -201,15 +203,15 @@ func TestRemove(t *testing.T) {
 
 	val, ok := cache.Get("foo")
 	assert.False(t, ok)
-	assert.Nil(t, val)
+	assert.Zero(t, val)
 }
 
 func TestEvictOldest(t *testing.T) {
 	var eviction bool
 
-	cache := New(Config{
+	cache := New(Config[string, string]{
 		Capacity: 1,
-		OnEviction: func(key, value interface{}) {
+		OnEviction: func(key, value string) {
 			eviction = true
 		},
 	})
@@ -222,7 +224,7 @@ func TestEvictOldest(t *testing.T) {
 
 	val, ok := cache.Get("foo")
 	assert.False(t, ok)
-	assert.Nil(t, val)
+	assert.Zero(t, val)
 
 	eviction = false
 	ok = cache.EvictOldest()
@@ -231,7 +233,7 @@ func TestEvictOldest(t *testing.T) {
 }
 
 func TestLen(t *testing.T) {
-	cache := New(Config{Capacity: 10})
+	cache := New(Config[int, int]{Capacity: 10})
 	for i := 0; i <= 9; i++ {
 		evict := cache.Set(i, i)
 		assert.False(t, evict)
@@ -241,7 +243,7 @@ func TestLen(t *testing.T) {
 }
 
 func TestClear(t *testing.T) {
-	cache := New(Config{Capacity: 10})
+	cache := New(Config[int, int]{Capacity: 10})
 	for i := 0; i <= 9; i++ {
 		evict := cache.Set(i, i)
 		assert.False(t, evict)
@@ -257,13 +259,13 @@ func TestClear(t *testing.T) {
 }
 
 func TestKeys(t *testing.T) {
-	cache := New(Config{Capacity: 10})
+	cache := New(Config[string, int]{Capacity: 10})
 	cache.Set("foo", 1)
 	cache.Set("bar", 2)
 
-	// key order isn't guarenteed
+	// key order isn't guaranteed
 	keys := cache.Keys()
-	sortedKeys := []string{keys[0].(string), keys[1].(string)}
+	sortedKeys := []string{keys[0], keys[1]}
 	sort.Strings(sortedKeys)
 
 	assert.Equal(t, 2, len(sortedKeys))
@@ -272,7 +274,7 @@ func TestKeys(t *testing.T) {
 }
 
 func TestOrderedKeys(t *testing.T) {
-	cache := New(Config{Capacity: 10})
+	cache := New(Config[string, int]{Capacity: 10})
 	cache.Set("foo", 1)
 	cache.Set("bar", 2)
 
@@ -284,7 +286,7 @@ func TestOrderedKeys(t *testing.T) {
 }
 
 func TestSetMaxAge(t *testing.T) {
-	cache := New(Config{Capacity: 10})
+	cache := New(Config[string, int]{Capacity: 10})
 	err := cache.SetMaxAge(-1 * time.Hour)
 	assert.Error(t, err)
 
@@ -293,7 +295,7 @@ func TestSetMaxAge(t *testing.T) {
 }
 
 func TestSetMinAge(t *testing.T) {
-	cache := New(Config{Capacity: 10, MaxAge: time.Hour})
+	cache := New(Config[string, int]{Capacity: 10, MaxAge: time.Hour})
 	err := cache.SetMinAge(-1 * time.Hour)
 	assert.Error(t, err)
 
@@ -304,8 +306,8 @@ func TestSetMinAge(t *testing.T) {
 func TestOnEviction(t *testing.T) {
 	var eviction bool
 
-	cache := New(Config{Capacity: 1})
-	cache.OnEviction(func(key, value interface{}) {
+	cache := New(Config[string, int]{Capacity: 1})
+	cache.OnEviction(func(key string, value int) {
 		eviction = true
 	})
 
@@ -318,11 +320,11 @@ func TestOnEviction(t *testing.T) {
 func TestOnExpiration(t *testing.T) {
 	var expiration bool
 
-	cache := New(Config{
+	cache := New(Config[string, int]{
 		Capacity: 1,
 		MaxAge:   time.Millisecond,
 	})
-	cache.OnExpiration(func(key, value interface{}) {
+	cache.OnExpiration(func(key string, value int) {
 		expiration = true
 	})
 
@@ -336,13 +338,13 @@ func TestOnExpiration(t *testing.T) {
 func TestActiveExpiration(t *testing.T) {
 	invoked := make(chan bool)
 
-	cache := New(Config{
+	cache := New(Config[string, int]{
 		Capacity:       1,
 		MaxAge:         time.Millisecond,
 		ExpirationType: ActiveExpiration,
 	})
 
-	cache.OnExpiration(func(key, value interface{}) {
+	cache.OnExpiration(func(key string, value int) {
 		invoked <- true
 	})
 
@@ -355,7 +357,7 @@ func TestActiveExpiration(t *testing.T) {
 }
 
 func TestResize(t *testing.T) {
-	cache := New(Config{
+	cache := New(Config[string, int]{
 		Capacity: 2,
 	})
 	cache.Set("a", 1)
@@ -387,12 +389,12 @@ func TestResize(t *testing.T) {
 
 func TestStats(t *testing.T) {
 	t.Run("reports capacity", func(t *testing.T) {
-		cache := New(Config{Capacity: 100})
+		cache := New(Config[string, int]{Capacity: 100})
 		assert.Equal(t, int64(100), cache.Stats().Capacity)
 	})
 
 	t.Run("reports count", func(t *testing.T) {
-		cache := New(Config{Capacity: 100})
+		cache := New(Config[int, int]{Capacity: 100})
 		for i := 0; i < 10; i++ {
 			cache.Set(i, i)
 		}
@@ -400,7 +402,7 @@ func TestStats(t *testing.T) {
 	})
 
 	t.Run("increments sets", func(t *testing.T) {
-		cache := New(Config{Capacity: 100, MaxAge: time.Second})
+		cache := New(Config[string, string]{Capacity: 100, MaxAge: time.Second})
 		for i := 0; i < 10; i++ {
 			cache.Set("foo", "bar")
 		}
@@ -408,7 +410,7 @@ func TestStats(t *testing.T) {
 	})
 
 	t.Run("increments gets", func(t *testing.T) {
-		cache := New(Config{Capacity: 100, MaxAge: time.Second})
+		cache := New(Config[string, int]{Capacity: 100, MaxAge: time.Second})
 		for i := 0; i < 10; i++ {
 			cache.Get("foo")
 		}
@@ -416,20 +418,20 @@ func TestStats(t *testing.T) {
 	})
 
 	t.Run("increments hits", func(t *testing.T) {
-		cache := New(Config{Capacity: 100, MaxAge: time.Second})
+		cache := New(Config[string, string]{Capacity: 100, MaxAge: time.Second})
 		cache.Set("foo", "bar")
 		cache.Get("foo")
 		assert.Equal(t, int64(1), cache.Stats().Gets)
 	})
 
 	t.Run("increments misses", func(t *testing.T) {
-		cache := New(Config{Capacity: 100, MaxAge: time.Second})
+		cache := New(Config[string, string]{Capacity: 100, MaxAge: time.Second})
 		cache.Get("foo")
 		assert.Equal(t, int64(1), cache.Stats().Misses)
 	})
 
 	t.Run("increments evictions", func(t *testing.T) {
-		cache := New(Config{Capacity: 1, MaxAge: time.Second})
+		cache := New(Config[int, int]{Capacity: 1, MaxAge: time.Second})
 		for i := 0; i < 10; i++ {
 			cache.Set(i, i)
 		}
@@ -437,7 +439,7 @@ func TestStats(t *testing.T) {
 	})
 
 	t.Run("delta stats", func(t *testing.T) {
-		cache := New(Config{Capacity: 100, MaxAge: time.Second})
+		cache := New(Config[string, string]{Capacity: 100, MaxAge: time.Second})
 		cache.Set("a", "1")
 		prev := cache.Stats()
 
@@ -460,7 +462,7 @@ func TestStats(t *testing.T) {
 	})
 
 	t.Run("copy", func(t *testing.T) {
-		cache := New(Config{Capacity: 100, MaxAge: time.Second})
+		cache := New(Config[string, string]{Capacity: 100, MaxAge: time.Second})
 		stats := cache.Stats()
 		stats.Hits++
 		stats.Misses++
@@ -470,7 +472,7 @@ func TestStats(t *testing.T) {
 }
 
 func BenchmarkCache(b *testing.B) {
-	cache := New(Config{Capacity: 100, MaxAge: time.Second})
+	cache := New(Config[string, string]{Capacity: 100, MaxAge: time.Second})
 
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {