This converts `agecache` to use Generics introduced in go1.18 for the `key` and `value`. It is a breaking change.

The advantages are:
- type safety guarantees
- better performance

before:
```go
go test --bench=. --benchmem
goos: darwin
goarch: amd64
pkg: github.com/deankarn/agecache
cpu: VirtualApple @ 2.50GHz
BenchmarkCache-8         2229212               537.5 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/deankarn/agecache    2.400s
```

after:
```go
go test --bench=. --benchmem
goos: darwin
goarch: amd64
pkg: github.com/deankarn/agecache
cpu: VirtualApple @ 2.50GHz
BenchmarkCache-8         2467549               483.2 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/deankarn/agecache    2.349s
```