@deankarn 2023-04-30
It appears that this package is no longer being maintained or maintainers are happy with the current implementation, in either case I'll be closing this PR.

For anyone interested or still seeking a generics based cache I have since created the following `lru` & `lfu` generic cache implementations.